+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## [1.2.0]
-- feature: default height & width
+- feature: default height & width (plugin configuration)
+- feature: open edit mask after video selection
 
 ## [1.1.0]
 - ILIAS 5.4 support